Fix: Translation collision
We have a dedicated file for translations that we are using exclusively
during the tests so that the tests do not rely on production data.

But, by construction we used the same translation keys and the test
translation file would erase the production one and be displayed on the
website.

On the other hand, some times, the production translation would erase
the test_translations leading to fleaky tests.

This is now fixed, all the tests fixtures are prefixed with `test_` and
this has been also impacted on all the tests checks