Added CapactitesTechniquesProfessionnellesEffectifsCVIntervenant component
lots of thing hapening here, I wanted to split in clear separate commit
but ended up in a mess so everything is here.

We've been using the stimulus components for reapeatable items
We've defined a repeatable concern for future repeatable fields
We've simplified the management of strong parameters because it was
already done in the model layer.